requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.9.0.post1
async-timeout==4.0.1
attrs==21.2.0
Babel==2.9.1
backports.zoneinfo==0.2.1
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==2.0.8
click==7.1.2
dnacentersdk==2.0.2
dnspython==2.2.0
fastjsonschema==2.15.1
Flask==1.1.2
Flask-PyMongo==2.3.0
frozenlist==1.2.0
future==0.18.2
geojson==2.5.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
meraki==1.15.0
multidict==5.2.0
numpy==1.21.4
pandas==1.3.4
PyJWT==1.7.1
pymongo==4.0.1
python-dateutil==2.8.2
python-dotenv==0.14.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.24.0
requests-toolbelt==0.9.1
six==1.16.0
tablib==3.1.0
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.25.10
webexteamssdk==1.6
Werkzeug==1.0.1
yarl==1.7.2